[Annotation] Save the comment changes made in reading mode (bug 1987427)

+++ src/display/annotation_storage.js
@@ -31,6 +31,8 @@ class AnnotationStorage {
 
   #modifiedIds = null;
 
+  #editorsMap = null;
+
   #storage = new Map();
 
   constructor() {
@@ -83,6 +85,13 @@ class AnnotationStorage {
    * @param {string} key
    */
   remove(key) {
+    const storedValue = this.#storage.get(key);
+    if (storedValue === undefined) {
+      return;
+    }
+    if (storedValue instanceof AnnotationEditor) {
+      this.#editorsMap.delete(storedValue.annotationElementId);
+    }
     this.#storage.delete(key);
 
     if (this.#storage.size === 0) {
@@ -122,11 +131,11 @@ class AnnotationStorage {
       this.#setModified();
     }
 
-    if (
-      value instanceof AnnotationEditor &&
-      typeof this.onAnnotationEditor === "function"
-    ) {
-      this.onAnnotationEditor(value.constructor._type);
+    if (value instanceof AnnotationEditor) {
+      (this.#editorsMap ||= new Map()).set(value.annotationElementId, value);
+      if (typeof this.onAnnotationEditor === "function") {
+        this.onAnnotationEditor(value.constructor._type);
+      }
     }
   }
 
@@ -250,6 +259,15 @@ class AnnotationStorage {
     this.#modifiedIds = null;
   }
 
+  updateEditor(annotationId, data) {
+    const value = this.#editorsMap?.get(annotationId);
+    if (value) {
+      value.updateFromAnnotationLayer(data);
+      return true;
+    }
+    return false;
+  }
+
   /**
    * @returns {{ids: Set<string>, hash: string}}
    */
@@ -258,15 +276,13 @@ class AnnotationStorage {
       return this.#modifiedIds;
     }
     const ids = [];
-    for (const value of this.#storage.values()) {
-      if (
-        !(value instanceof AnnotationEditor) ||
-        !value.annotationElementId ||
-        !value.serialize()
-      ) {
-        continue;
+    if (this.#editorsMap) {
+      for (const value of this.#editorsMap.values()) {
+        if (!value.serialize()) {
+          continue;
+        }
+        ids.push(value.annotationElementId);
       }
-      ids.push(value.annotationElementId);
     }
     return (this.#modifiedIds = {
       ids: new Set(ids),
